requirements.txt
pandas
duckdb
requests
requests-cache
python-dotenv
seaborn
matplotlib
numpy
pandas